refactor: 모델 프로토콜 파일 분리
